ExpressLRS - Change the failsafe missed packets threshold to 2000.
Likely this needs more work, but this change is required for Low LQ
testing.

+++ src/main/rx/expresslrs_impl.h
@@ -49,7 +49,7 @@ typedef struct elrsReceiver_s {
     uint8_t uplinkLQ;
 
     uint32_t lastValidPacketUs;
-    uint8_t missedPackets;
+    uint16_t missedPackets;
 
 
     uint16_t cycleInterval;
